+++ tests/modeldiffs/vanilla_sgd_pytorch.py
@@ -1,0 +1,23 @@
+import torch
+
+from algorithmic_efficiency import spec
+from reference_algorithms.target_setting_algorithms.data_selection import \
+    data_selection
+from reference_algorithms.target_setting_algorithms.pytorch_submission_base import \
+    update_params 
+
+def init_optimizer_state(workload: spec.Workload,
+                           model_params: spec.ParameterContainer,
+                           model_state: spec.ModelAuxiliaryState,
+                           hyperparameters: spec.Hyperparameters,
+                           rng: spec.RandomState) -> spec.OptimizerState:
+  """Creates a Vanilla SGD Optimizer."""
+  del model_state
+  del rng
+
+  optimizer_state = {
+      'optimizer':
+          torch.optim.SGD(model_params.parameters(), lr=0.001, weight_decay=0)
+  }
+
+  return optimizer_state
